separate sqlite impl from header
