+++ src/geopackage/sqlite/iterator.cpp
@@ -1,0 +1,118 @@
+#include "SQLite.hpp"
+
+using namespace geopackage;
+
+sqlite_iter::sqlite_iter(stmt_t stmt)
+  : stmt(stmt)
+{
+    this->column_count = sqlite3_column_count(this->ptr());
+    this->column_names = std::vector<std::string>();
+    this->column_names.reserve(this->column_count);
+    this->column_types = std::vector<int>();
+    this->column_types.reserve(this->column_count);
+
+    for (int i = 0; i < this->column_count; i++) {
+        this->column_names.push_back(sqlite3_column_name(this->ptr(), i));
+        this->column_types.push_back(sqlite3_column_type(this->ptr(), i));
+    }
+};
+
+sqlite3_stmt* sqlite_iter::ptr() const noexcept
+{
+    return this->stmt.get();
+}
+
+void sqlite_iter::handle_get_index(int col) const
+{
+    if (this->done()) {
+        throw sqlite_get_done_error;
+    }
+
+    if (this->current_row() == -1) {
+        throw sqlite_get_notstarted_error;
+    }
+
+    if (col < 0 || col >= this->column_count) {
+        throw std::out_of_range(
+            "column " + std::to_string(col) + " out of range of " + std::to_string(this->column_count) + " columns"
+        );
+    }
+}
+
+bool sqlite_iter::done() const noexcept
+{
+    return this->iteration_finished;
+}
+
+sqlite_iter& sqlite_iter::next()
+{
+    if (!this->done()) {
+        const int returncode = sqlite3_step(this->ptr());
+        if (returncode == SQLITE_DONE) {
+            this->iteration_finished = true;
+        }
+        this->iteration_step++;
+    }
+
+    return *this;
+}
+
+sqlite_iter& sqlite_iter::restart()
+{
+    sqlite3_reset(this->ptr());
+    this->iteration_step = -1;
+    this->iteration_finished = false;
+    return *this;
+}
+
+void sqlite_iter::close()
+{
+    this->~sqlite_iter();
+}
+
+int sqlite_iter::current_row() const noexcept
+{
+    return this->iteration_step;
+}
+
+int sqlite_iter::num_columns() const noexcept
+{
+    return this->column_count;
+}
+
+int sqlite_iter::column_index(const std::string& name) const noexcept
+{
+    const ptrdiff_t pos =
+      std::distance(this->column_names.begin(), std::find(this->column_names.begin(), this->column_names.end(), name));
+
+    return pos >= this->column_names.size() ? -1 : pos;
+}
+
+template<>
+std::vector<uint8_t> sqlite_iter::get<std::vector<uint8_t>>(int col) const
+{
+    this->handle_get_index(col);
+    return *reinterpret_cast<const std::vector<uint8_t>*>(sqlite3_column_blob(this->ptr(), col));
+}
+
+template<>
+double sqlite_iter::get<double>(int col) const
+{
+    this->handle_get_index(col);
+    return sqlite3_column_double(this->ptr(), col);
+}
+
+template<>
+int sqlite_iter::get<int>(int col) const
+{
+    this->handle_get_index(col);
+    return sqlite3_column_int(this->ptr(), col);
+}
+
+template<>
+std::string sqlite_iter::get<std::string>(int col) const
+{
+    this->handle_get_index(col);
+    // TODO: this won't work with non-ASCII text
+    return std::string(reinterpret_cast<const char*>(sqlite3_column_text(this->ptr(), col)));
+}
